requirements.txt
# Flask web framework
flask==3.0.2
flask-cors==4.0.0

# Face recognition
face_recognition==1.3.0
dlib==19.24.0

# Image processing
Pillow==10.2.0
numpy==1.26.4

# HTTP requests
requests==2.31.0 